requirements.txt
alembic==1.14.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lazy_loader==0.4
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
pycparser==2.22
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
ratelimit==2.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.1
scikit-image==0.25.0
scipy==1.15.0
setuptools==75.7.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.37
sympy==1.13.1
tifffile==2024.12.12
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
